requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.48
Flask-WTF==1.0.0
python-dotenv==0.19.0
geopy==2.2.0